+++ modules/books-constructor.js
@@ -1,0 +1,7 @@
+export default class Book {
+  constructor(title, author, id) {
+    this.id = id;
+    this.title = title;
+    this.author = author;
+  }
+}
